Leave terraform at version 0.11.11
There seems to be some compatibility problems between newer versions of
Terraform and the kubespray openstack templates we use.  The problems
might be due to unpinned plugin versions (the ones downloaded by
terraform when it performs `init`, in which case it wouldn't be solved
by avoiding this update to Terraform 0.11.14) but I'm not sure.